Move networkSelector under login screen
