+++ src/components/screens/login/networkSelector/networkSelector.js
@@ -5,14 +5,14 @@ import liskClient from 'Utils/lisk-client'; // eslint-disable-line
 import { withTranslation } from 'react-i18next';
 import { withRouter } from 'react-router';
 import { to } from 'await-to-js';
-import { PrimaryButton, SecondaryButton } from '../../../../toolbox/buttons';
-import { Input } from '../../../../toolbox/inputs';
-import { addHttp, getAutoLogInData, findMatchingLoginNetwork } from '../../../../../utils/login';
-import getNetwork, { getNetworksList } from '../../../../../utils/getNetwork';
-import networks from '../../../../../constants/networks';
+import { PrimaryButton, SecondaryButton } from '../../../toolbox/buttons';
+import { Input } from '../../../toolbox/inputs';
+import { addHttp, getAutoLogInData, findMatchingLoginNetwork } from '../../../../utils/login';
+import getNetwork, { getNetworksList } from '../../../../utils/getNetwork';
+import networks from '../../../../constants/networks';
 import styles from './networkSelector.css';
-import keyCodes from '../../../../../constants/keyCodes';
-import DropdownButton from '../../../../toolbox/dropdownButton';
+import keyCodes from '../../../../constants/keyCodes';
+import DropdownButton from '../../../toolbox/dropdownButton';
 
 class NetworkSelector extends React.Component {
   // eslint-disable-next-line max-statements
