+++ src/components/screens/login/networkSelector/networkSelector.test.js
